fix(npm): Serenity/JS compiles to ES5 instead of ES6 to fix issues with the module loader
To use Serenity/JS on a node.js project, the library has to be compiled to ES5. This is required,

because: 1) Node.js doesn't support the ES6 module loading mechanism 2) Even when the TS compiler is

configured to generate ES5-compatible module declarations, it still seems to be using ES6.