feat(integration): Stand-ins enable taking screenshots when there are no Actors on the Stage.
The Photographer relies on Actors being present on the Stage to recognise whose activities should
be captured.

A Stand-in substitutes for the Actor if they're missing, allowing you to capture rich, illustrated reports
of your acceptance test scenarios even if you haven't (yet) adopted the Screenplay Pattern, nor have defined
the actors.

This functionality is useful when Serenity/JS is introduced to an existing code base as it allows
for gradual and safe adoption of the library.

Thanks to @awashbrook for the "Stand-in" metaphor!