feat(serenity-protractor): Interactions: Execute.asyncScript(), Execute.script() and Evaluate.script()
New interactions help to apply changes to the internal state of an Angular application.

This is not something you should probably be using when writing a regular acceptance test,
but can come in handy when you need to create a "fast-track" equivalent of a slow UI-based task.

More about the "fast-track" pattern shortly at https://janmolak.com