fix(cucumber,mocha): the stageCue timeout is configurable
affects: serenity-js

Protractor requires Serenity/JS to notify it as soon as a test scenario is completed, which means
that Serenity/JS needs to wait for any outstanding tasks, such as saving screenshots to disk, before
it can do that.

The previous timeout of this operation was hard-coded to 10s, but when Serenity/JS
is used with BrowserStack, SauceLabs or any other remote Selenium grid, it might take a bit longer
for the screenshots to be downloaded, which means that the timeout was not sufficient.

This change introduces a new configurable property, which can be specified in Protractor config:

```
import { Duration } from 'serenity-js/lib/serenity/duration';

{
  serenity: {
    timeouts: {
      stageCue: Duration.ofSeconds(s: number)  // or Duration.ofMillis(ms: number)
    }
  }
}
```

The `stageCue` timeout defaults to 30s maximum.

ISSUES CLOSED: #34