fix(serenity-bdd): reports are correctly stored to disk
SerenityBDDReporter used to generate the reports upon a TestRunFinished event, which is too late in
the process and doesn't leave enough time for them to be saved by the ArtifactArchiver. Introducing
a new event TestRunFinishes that occurs before TestRunFinished and generating the reports upon it
helps to synchronise the storage correctly. Test runner adapters now have to: 1) emit
TestRunFinishes when the test suite is about to complete, 2) wait for the next cue, 3) emit
TestRunFinished