feat(core): discardable Abilities and shared notes
- Abilities implementing the new Discardable interface are automatically discarded after every
scenario. This allows them to free up any resources they're no longer using, such as a database or a
websocket connection, for example. | The ability to TakeNotes.usingASharedNotepad can be used to
share notes between the actors. This is useful when you have one actor start up the local test
server, and the other one navigate to it in their browser. Please see the API docs for examples.