fix(protractor): photographer will not try to capture the screenshot if the actor has no browser
In multi-actor scenarios it is possible to have an actor with no BrowseTheWeb ability set up the
test environment before another actor interacts with it via their web browser. If in such scenario
the Photographer is configured to TakePhotosOfInteractions, it will no longer attempt to capture
screenshots of an actor who doesn't have the ability to BrowseTheWeb, and therefore doesn't have
access to the browser.