feat(serenity-bdd): upgraded to Serenity BDD CLI 2.1.12 and improved the downloader
serenity-bdd downloader will now inform you where it's downloading the jar from/to so that it's easier to debug any network-related issues.
The downloader will also verify the checksum of the downloaded file to ensure that it's not corrupted.

Added instructions on how to configure Serenity BDD reporting CLI - closes #625