fix(cucumber): don't pass the "rerun" file to Cucumber if it doesn't exist
When executing Cucumber in re-run mode, specifying a path to a "rerun" file that didn't exist
resulted in a Cucumber throwing an error. This was problematic as it required the user to modify
Cucumber config each time they wanted to re-run failed tests. This change prevents Serenity from
passing the location of "rerun" file to Cucumber if it doesn't exist, allowing the tests to be
executed normally if there were no failures, or re-ran otherwise

fix #971 re protractor-cucumber-framework/protractor-cucumber-framework#219