feat(core): replaced `Adapter` with `QuestionAdapter` and introduced `Optional`
`QuestionAdapter` is a far more powerful alternative to the previous Adapter implementation.
It is instantiated upon a call to `Question.about` and creates a Proxy around the returned object such that the fields
of the object can be used as Questions, and their methods as either Questions or Interactions.
Additionally, `Optional` introduces a new method `isPresent()` which allows the developer to check if all the fields in a given chain
are defined;

For example:

```typescript
const value = () => Question.about<{ a?: { b?: { c?: string }}}>('some value', actor => {
  return {
    a: {
      b: {
        // no "c"
      }
    }
});

actor.attemptsTo(
  Ensure.that(value().a.b.c.isDefined(), equals(false)),
  Ensure.that(value().a.b.c, equals(undefined)),
)
```

RE #1103