feat(web): combined Frame and PageElement so they can be Switch-ed to
Interaction to `Switch.to(switchable)` works for both `Page` and `PageElement` now. For
`PageElement`, if the element represents a `frame` or `iframe`, `Switch.to(element)` switches the
top-level browsing context to that frame. When a `PageElement` represents any other HTML element,
invoking `Switch.to(element)` sets browser focus on that element.

re #82 #227 #233 #365