fix(web): improved JavaScript dialog-related error handling when taking screenshots
Since attempting to take a screenshot when a dialog is present leads to Puppeteer hanging (see
puppeteer/puppeteer#2481), Photographer used to check if a dialog was present before taking a
screenshot. However, since the Webdriver protocol doesn't provide any way for us to check for an
alert other than trying to switch to it, attempting to switch to an alert that wasn't there would
lead to errors with Webdriver protocol-based integration tools (Protractor, WebdriverIO). This
change improves the behaviour of the Photographer and moves the error handling logic to
tool-specific Page implementations.

fix #1156