feat(web): PageElement-releated expectations now also check if the element is present
Expectations that a PageElement isActive, isClickable, isEnabled, isSelected, or isVisible will now assume that if the element is not present, it cannot therefore be active, clickable, enabled, selected, or visible, respectively.
This should be reasonable and intuitive behaviour in most cases. However, note that if in your scenario you want to ensure that the element is, for example, "not enabled", and there is a possibility for the element to not be there at all, you might want to compose expectations as follows:

```
Ensure.that(element, and(isPresent(), not(isEnabled()))
```

This way you prevent an error that might be caused by `not(isEnabled())` returning `true` when the element is also not present.

Another thing to note is that this change will make it easier for you to perform expectations on PageElements that might not return an element.

For example, the following code will now work:
```
const listItems = PageElements.located(By.css('li'))

Wait.until(listItems.first(), isVisible())
```

Before this change, you had to make sure you checked if the element was also present to avoid the test throwing an error when there were no elements returned:
```
const listItems = PageElements.located(By.css('li'))

Wait.until(listItems.first(), and(isPresent(), isVisible()))
```

closes #1255