fix(playwright): corrected not(isPresent()) for PlaywrightPageElement
previously, isPresent() would try to retrieve the element waiting for it to be attached in the DOM.
However, this caused not(isPresent()) to time out waiting for the element to show up. This change
corrects this behaviour, making not(isPresent())) terminate early if the element is not present.

Related tickets: re #1240