feat(console-reporter): improved support for tests executed in parallel
ConsoleReporter used to print to the terminal as soon as it received a test-related event from the
test runner. With the advent of test runners executing tests in parallel, such as Cucumber 8 and
Playwright Test this became problematic as sometimes information coming from different test
scenarios would overlap. This change makes the ConsoleReporter print to the terminal only when the
test scenario is finished and lays foundation for further improvements.

Related tickets: re #1240