fix(web): interaction to Clear.theValueOf(field) triggers compatible events for all web tools
Selenium WebDriver, WebdriverIO with DevTools, and Playwright each
have a slightly different way of performing the operation to clear the value of an input field
or a "contenteditable" field. In particular, some of those implementations trigger the 'blur' event,
some don't. Some trigger a 'keydown' event, some won't.

Those subtle incompatibilities could result in differerent behaviour of the system under test,
especially when it relies on human interaction with the UI to result
in certain events being triggered. E.g. it has an onBlur event handler that triggers
field validation of a form.

This fix ensures that all the implementations supported by Serenity/JS, so:
- Selenium 3 via Protractor,
- WebDriver protocol via WebDriverIO,
- DevTools protocol via WebDriverIO,
- DevTools protocol via Playwright
behave in a compatible way, so that they:
- issue 'keydown' events with Delete key to clear the field
- don't trigger the 'blur' event

This way the interaction to Clear can be followed by the interaction to Enter or Press
while the input maintains its focus

Closes #1329