feat(core): better assertion errors reporting in Visual Studio Code
The message of a Serenity/JS AssertionError now includes the diff of expected and actual values,
which allows for better error reporting in Visual Studio Code. I've also rewritten the diff
mechanism itself to make it easier for developers to read, and to allow support for colour-coding in
the future.

Related tickets: Closes #1486