fix(core): further improvements to how the expected vs received values are rendered in VS Code
VSCode with Playwright Test plugin doesn't use fixed width fonts to display the error message. This
makes spotting the difference between the expected vs actual values much harder than it could've
been. With this patch, I've renamed "actual" to "received" so that the meaning is not only clearer,
but the layout is improved as well since "received" has the same number of characters as "expected"

Related tickets: re #1486