fix(playwright-test): default to using file name as feature name when describe blocks are absent
In test runners like Mocha or Jasmine, test scenarios defined outside of a `describe` block are
considered global and are generally discouraged.
However, Playwright Test avoids this problem by scoping such test scenarios to file in which
they are defined.

This change ensures that when no `describe` blocks surround a given Playwright Test scenario,
the name of its file is used as the name of the feature.

Related tickets: closes #1495