fix(serenity-bdd): prevent invalid Serenity BDD JSON reports from being sent to processing
Serenity BDD CLI requires the `name` field of the JSON report to be populated, otherwise it produces
an empty HTML report. While this is something Serenity BDD should complain about during processing,
Serenity/JS should not allow for invalid reports to be produced in the first place. This change
addresses that.

Related tickets: re #1495