fix(rest): correctly resolve relative paths in REST requests
Axios seems to have a bug that manifests itself when the baseURL is defined with a path, e.g.
https://example.org/api/v1
Then, when an individual request is supposed to be sent to an absolute
path, e.g. /secure/oauth, Axios incorrectly resolves this combination as https://example.org/api/v1/secure/oauth
instead of https://example.org/secure/oauth
This change fixes this behaviour by relying on Nodejs built-in URL resolution: new URL(path, baseURL)