fix(playwright): ensure ExecuteScript runs in the context of the currently active iframe
Previously, Serenity/JS Playwright integration would execute the script provided via the interaction
to ExecuteScript in the context of the top-level frame, instead of the frame activated by the
interaction to Switch.to(frame). With this change, interaction to Switch.to(frame) temporarily
changes the injected JavaScript execution context to the active frame.