feat(core): introduced specialised events to mark actors' exit from the stage
ActorStageExitAttempted event includes a serialised actor and its abilities allowing reporters to
include more information about the actors in test reports.
ActorStageExitCompleted is emitted when the actor is correctly released after the test scenario/test suite is finished (depending on the scope of the actor)
ActorStageExitFailed is emitted if releasing the actor resulted in an error.