feat(playwright): refactored SerenityFixtures, corrected BrowseTheWebWithPlaywright parameters
- The ability to BrowseTheWebWithPlaywright now separates the BrowserContextOptions parameter,
available when creating an ability using the Playwright browser but not the page, from the
ExtraBrowserContextOptions, available for both the browser and the page. This change tightens the
API and clarifies that certain options are not available when passing an already-configured page.

- SerenityOptions and SerenityFixtures are now merged and available as SerenityFixtures to be used
with defineConfig\<SerenityFixtures\> to follow the current Playwright Test convention

- New type SerenityWorkerFixtures introduced to allow for separating test-scoped and worker-scoped
Serenity fixtures in future releases